+++ python/import/log.py
@@ -1,0 +1,4 @@
+#!/bin/env python
+
+def log_error(message):
+    print("ERROR: %s" % message)