+++ python/import/main.py
@@ -1,0 +1,4 @@
+#!/bin/env python
+from log import log_error
+
+log_error("this is a error log")